chore: move nodebootstrapping logic to separate file, remove the need to pass down func param
